remove compiler warning

+++ six/modules/c++/six.sicd/source/ComplexToAMP8IPHS8I_unseq.cpp
@@ -857,7 +857,7 @@ struct AMP8I_PHS8I_array final : public std::array<AMP8I_PHS8I, N>
     AMP8I_PHS8I_array& operator=(const AMP8I_PHS8I_unseq<IntV>&) = delete; // should only be using move-assignment
     AMP8I_PHS8I_array& operator=(AMP8I_PHS8I_unseq<IntV>&& other)
     {
-        for (int i = 0; i < N; i++)
+        for (int i = 0; i < gsl::narrow<int>(N); i++)
         {
             (*this)[i].phase = gsl::narrow<uint8_t>(other.phase[i]);
             (*this)[i].amplitude = gsl::narrow<uint8_t>(other.amplitude[i]);
